refactor(transaction): Use LocalDate in TransactionUiState and handle DatePickerDialog
