+++ presentation/src/commonMain/kotlin/com/moneyplusplus/presentation/feature/transaction/TransactionIntent.kt
@@ -9,6 +9,7 @@ sealed class TransactionIntent : UiIntent {
     data object OnFilterSheetDismissed : TransactionIntent()
     data object OnDateClick : TransactionIntent()
     data class OnDateSelected(val date: LocalDate) : TransactionIntent()
+    data object OnDatePickerDialogDismiss : TransactionIntent()
     data class OnTransactionTypeClick(val type: TransactionTypeFilter) : TransactionIntent()
     data object OnAddTransactionClick : TransactionIntent()
     data object OnAddTransactionSheetDismissed : TransactionIntent()
