+++ presentation/src/commonMain/kotlin/com/moneyplusplus/presentation/feature/transaction/TransactionUiState.kt
@@ -3,6 +3,7 @@ package com.moneyplusplus.presentation.feature.transaction
 import com.moneyplusplus.presentation.base.UiState
 import com.moneyplusplus.presentation.model.CategoryUiModel
 import com.moneyplusplus.presentation.model.TransactionUiModel
+import kotlinx.datetime.LocalDate
 
 data class TransactionUiState(
     val transactions: List<TransactionUiModel> = emptyList(),
@@ -10,10 +11,10 @@ data class TransactionUiState(
     val isError: Boolean = false,
     val errorMessage: String? = null,
     val typeFilter: TransactionTypeFilter = TransactionTypeFilter.ALL,
-    val date: String = "",
+    val date: LocalDate? = null,
     val showCategoriesFilterBottomSheet: Boolean = false,
     val showAddTransactionBottomSheet: Boolean = false,
     val categories: List<CategoryUiModel> = emptyList(),
     val selectedCategoryIds: List<String> = emptyList(),
-    val showDatePicker: Boolean = false
+    val showDatePickerDialog: Boolean = false
 ) : UiState
